Move layers list view to _qt
